requirements.txt
boto3>=1.26.0
flask>=2.0.0
psycopg2-binary>=2.9.0
pyyaml>=6.0.0
requests>=2.28.0
